Restructure screens structure
