+++ App.tsx
@@ -3,7 +3,7 @@ import { PescaContextProvider } from '@moneyboy/context/PescaContext';
 import { SettingsContextProvider } from '@moneyboy/context/SettingsContext';
 import { StorageContextProvider } from '@moneyboy/context/StorageContext';
 import { StyleContextProvider } from '@moneyboy/context/StyleContext';
-import { AppContainer } from '@moneyboy/views/AppContainer';
+import { AppContainer } from '@moneyboy/screens/AppContainer';
 import React from 'react';
 import { StatusBar } from 'react-native';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
