+++ src/screens/AppContainer.tsx
@@ -1,13 +1,13 @@
 import { PescaTabBar } from '@moneyboy/components/navigation/PescaTabBar';
 import { AuthContext } from '@moneyboy/context/AuthContext';
 import { StyleContext } from '@moneyboy/context/StyleContext';
-import { EventView } from '@moneyboy/views/pages/EventView';
-import { GroupView } from '@moneyboy/views/pages/GroupView';
-import { HistoryView } from '@moneyboy/views/pages/HistoryView';
-import { LoginView } from '@moneyboy/views/pages/LoginView';
-import { MainView } from '@moneyboy/views/pages/MainView';
-import { RegisterView } from '@moneyboy/views/pages/RegisterView';
-import SettingsPanel from '@moneyboy/views/SettingsPanel';
+import { LoginView } from '@moneyboy/screens/auth/login';
+import { RegisterView } from '@moneyboy/screens/auth/register';
+import { EventView } from '@moneyboy/screens/events';
+import { GroupView } from '@moneyboy/screens/groups';
+import { HistoryView } from '@moneyboy/screens/history';
+import { MainView } from '@moneyboy/screens/main';
+import SettingsPanel from '@moneyboy/screens/SettingsPanel';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import { NavigationContainer } from '@react-navigation/native';
 import React, { useContext } from 'react';
